requirements.txt
# Core dependencies
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0

# Machine learning
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
yellowbrick>=1.4.0

# Configuration
pyyaml>=5.4.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Optional development tools
ipython>=7.0.0
jupyter>=1.0.0

